Implement lobby link and play area helpers
